still trying to fix stuff

+++ components/spotify/cspot/bell/main/io/include/BellSocket.h
@@ -10,8 +10,8 @@ class Socket {
 
   virtual void open(const std::string& host, uint16_t port) = 0;
   virtual size_t poll() = 0;
-  virtual size_t write(uint8_t* buf, size_t len) = 0;
-  virtual size_t read(uint8_t* buf, size_t len) = 0;
+  virtual int write(uint8_t* buf, size_t len) = 0;
+  virtual int read(uint8_t* buf, size_t len) = 0;
   virtual bool isOpen() = 0;
   virtual void close() = 0;
   virtual int getFd() = 0;
