+++ components/spotify/cspot/bell/main/io/include/TCPSocket.h
@@ -85,12 +85,12 @@ class TCPSocket : public bell::Socket {
     isClosed = false;
   }
 
-  size_t read(uint8_t* buf, size_t len) {
-    return recv(sockFd, (char*)buf, len, 0);
+  int read(uint8_t* buf, size_t len) {
+    return (int)recv(sockFd, (char*)buf, len, 0);
   }
 
-  size_t write(uint8_t* buf, size_t len) {
-    return send(sockFd, (char*)buf, len, 0);
+  int write(uint8_t* buf, size_t len) {
+    return (int)send(sockFd, (char*)buf, len, 0);
   }
 
   size_t poll() {
